+++ src/app/components/token-template-selector/oneclient-in-oneprovider-template-hint.js
@@ -1,0 +1,11 @@
+/**
+ * Hint for oneclient token template for specific Oneprovider.
+ *
+ * @author Agnieszka Raczek
+ * @copyright (C) 2025 Onedata (onedata.org)
+ * @license This software is released under the MIT license cited in 'LICENSE.txt'.
+ */
+
+import TemplateHint from './template-hint';
+
+export default class OneclientInOneproviderTemplateHintComponent extends TemplateHint {}