+++ src/app/components/token-template-selector/readonly-data-for-user-template-hint.js
@@ -1,0 +1,11 @@
+/**
+ * Hint for read-only data token template for specific user.
+ *
+ * @author Agnieszka Raczek
+ * @copyright (C) 2025 Onedata (onedata.org)
+ * @license This software is released under the MIT license cited in 'LICENSE.txt'.
+ */
+
+import TemplateHint from './template-hint';
+
+export default class ReadonlyDataForUserTemplateHintComponent extends TemplateHint {}