+++ src/tests/integration/components/token-template-selector-test.js
@@ -12,6 +12,9 @@ const templates = [{
   name: 'oneclient',
   category: basicCategory,
 }, {
+  name: 's3',
+  category: basicCategory,
+}, {
   name: 'oneproviderRest',
   category: basicCategory,
 }, {